+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not released
 
 - Extend SelectField to cover more use cases [#796](https://github.com/CartoDB/carto-react/pull/796)
+- Show widget skeleton only once [#811](https://github.com/CartoDB/carto-react/pull/811)
 - Added filterable prop to TimeSeriesWidget [#808](https://github.com/CartoDB/carto-react/pull/808)
 - Fix dataSources store type [#807](https://github.com/CartoDB/carto-react/pull/807)